+++ packages/dodo-ui/src/lib/components/Form/Button/Button.svelte
@@ -52,10 +52,10 @@
 </script>
 
 <script lang="ts">
-  import type { ComponentSize } from '$lib/attributes/size.js';
-  import type { ComponentColor } from '$lib/attributes/color.js';
-  import type { ComponentRoundness } from '$lib/attributes/roundness.js';
-  import type { ComponentVariant } from '$lib/attributes/variant.js';
+  import type { ComponentSize } from '@flightlesslabs/dodo-core-attributes';
+  import type { ComponentColor } from '@flightlesslabs/dodo-core-attributes';
+  import type { ComponentRoundness } from '@flightlesslabs/dodo-core-attributes';
+  import type { ComponentVariant } from '@flightlesslabs/dodo-core-attributes';
   import type { ButtonAsAnchorProps as ButtonAsAnchorPropsBase } from './ButtonAsAnchor.svelte';
   import ButtonAsAnchor from './ButtonAsAnchor.svelte';
   import type { ButtonAsButtonProps as ButtonAsButtonPropsBase } from './ButtonAsButton.svelte';