+++ packages/dodo-ui/src/lib/components/Form/InputEnclosure/InputEnclosure.svelte
@@ -57,8 +57,8 @@
 </script>
 
 <script lang="ts">
-  import type { ComponentSize } from '$lib/attributes/size.js';
-  import type { ComponentRoundnessShape } from '$lib/attributes/roundness.js';
+  import type { ComponentSize } from '@flightlesslabs/dodo-core-attributes';
+  import type { ComponentRoundnessShape } from '@flightlesslabs/dodo-core-attributes';
   import InputAffix from './InputAffix.svelte';
 
   let {